Increased the heart's height
For some reason this was really hard. Please don't mess with it. The pixels are a static amount. Maybe in the future I can try to change that, but if it's ok I would prefer not to mess with this anymore.

Also, I added a float style to the inline code for the svg tag. That will migrate to the style tag at the top eventually.