Turned heart into a loading bar
CSS and Javascript files have been attached. Added a small amount of styling to the heart. Replaced svg tag with in-line div attributes. Corrected the drawBar function a little. Everything is not up and working completely, but it is all connected.

Also, for some reason the heart's path is cut off a little at the bottom. I don't know what to do about that. It's probably fine.